requirements.txt
dlib==19.19.0
Click==7.0
numpy==1.18.1
face_recognition_models==0.3.0
opencv_python==4.2.0.32
requests==2.23.0
imutils==0.5.3
Pillow==7.1.1
